LayoutTests/http/tests/contentextensions/frame-url-expected.txt

CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/frame-url.html from loading a resource from http://127.0.0.1:8000/contentextensions/resources/iframe.html
CONSOLE MESSAGE: Resource blocked by content blocker
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8000/contentextensions/resources/iframe.html due to access control checks.
ALERT: fetch was blocked
